requirements.txt
numpy
pandas 
regex 
psycopg2-binary
fastapi
uvicorn 
python-dotenv
python-multipart